requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
dataclasses-json==0.6.7
distro==1.9.0
Django==5.1.7
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
Faker==37.1.0
frozenlist==1.5.0
greenlet==3.1.1
groq==0.20.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.21
langchain-community==0.3.20
langchain-core==0.3.49
langchain-groq==0.3.1
langchain-text-splitters==0.3.7
langsmith==0.3.19
marshmallow==3.26.1
multidict==6.2.0
mypy-extensions==1.0.0
numpy==2.2.4
orjson==3.10.16
packaging==24.2
propcache==0.3.1
psycopg2-binary==2.9.10
pydantic==2.11.1
pydantic-settings==2.8.1
pydantic_core==2.33.0
PyJWT==2.9.0
pypdf==5.4.0
python-dotenv==1.1.0
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
sniffio==1.3.1
SQLAlchemy==2.0.40
sqlparse==0.5.3
tenacity==9.0.0
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.0
tzdata==2025.2
urllib3==2.3.0
yarl==1.18.3
zstandard==0.23.0
